% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Firedam.csv data from textbook with cost of fire damage in thousands
of dollars and distance of home to nearest fire station}
\format{
An object of class \code{data.frame} with 15 rows and 2 columns.
}
\usage{
fire
}
\description{
Firedam.csv data from textbook with cost of fire damage in thousands
of dollars and distance of home to nearest fire station
}
\keyword{datasets}
