% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial experiment simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a barplot of the simulated binomial experiment with given parameters
}
\description{
Binomial experiment simulation
}
\examples{
mybin(iter = 1000, n = 20, p = .1)

}
