% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootsrap interval function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample data}

\item{fun}{function to apply}

\item{alpha}{significance level of confidence linterval}

\item{cx}{text size}

\item{...}{aditional arguments}
}
\value{
A histogram of boostrap sample statistics
}
\description{
Bootsrap interval function
}
\examples{
myboot2(iter = 100000, x = rnorm(25,mean=25,sd=10), fun = "mean", alpha = 0.05)

}
