% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt function}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{size of sample}

\item{iter}{number of trials}
}
\value{
a histogram of uniformed distribution generated from the parameters
}
\description{
myclt function
}
\examples{
myclt(3,1000)

}
