% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Generate a normal curve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{probability limit}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
a graph of the normal curve using the function
}
\description{
Generate a normal curve
}
\examples{
myncurve(1, 0, 1)

}
