% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedline.R
\name{myplot}
\alias{myplot}
\title{Fitted line model for quad.lm}
\usage{
myplot(x)
}
\arguments{
\item{x}{- BHDiameter value}
}
\value{
predicted height value
}
\description{
Fitted line model for quad.lm
}
\examples{
myplot(3)

}
