% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on flight}

\item{gamma}{probability a plane wille be truly overbooked}

\item{p}{probability of a "show"}
}
\value{
A named list of nd, nc, N, p, and gamma
}
\description{
ntickets function
}
\examples{
ntickets(200, 0.02, 0.95)

}
